+++ semver-project-gradle-plugin/main/kotlin/com/javiersc/semver/project/gradle/plugin/tasks/PushSemverTagTask.kt
@@ -33,10 +33,11 @@ public abstract class PushSemverTagTask : DefaultTask() {
     }
 
     public companion object {
-        public const val TaskName: String = "pushSemverTag"
+
+        public const val NAME: String = "pushSemverTag"
 
         internal fun register(project: Project, gitTagBuildService: Provider<GitBuildService>) {
-            project.tasks.register<PushSemverTagTask>(TaskName).configure { task ->
+            project.tasks.register<PushSemverTagTask>(NAME).configure { task ->
                 task.tagPrefixProperty.set(project.tagPrefixProperty)
                 task.projectTagPrefix.set(project.projectTagPrefix)
                 task.version.set(project.version.toString())